fabrics: move nbft code to the library
Move more discovery code to the library.
